requirements_extract.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
scikit-learn>=0.24.0
scipy>=1.6.0
argparse
pathlib